requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-cosmos
azure-functions
azure-storage-blob
numpy<2.0
opencensus-ext-azure
opencensus-ext-logging
opencensus
pandas
scikit-learn
scipy
scikit-surprise